requirements.txt
absl-py==0.1.11
astor==0.6.2
bleach==1.5.0
click==6.7
coverage==4.5.1
cycler==0.10.0
decorator==4.2.1
Flask==0.12.2
Flask-Session==0.3.1
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.10
Keras==2.1.5
keras-vis==0.4.1
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.0
networkx==2.1
nose2==0.7.4
numpy==1.14.2
Pillow==5.0.0
protobuf==3.5.2
pyparsing==2.2.0
python-dateutil==2.7.0
pytz==2018.3
PyWavelets==0.5.2
PyYAML==3.12
scikit-image==0.13.1
scipy==1.0.0
six==1.11.0
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
Werkzeug==0.14.1
